% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{CASMAPenv}
\alias{CASMAPenv}
\title{Global variables environment}
\format{An object of class \code{environment} of length 3.}
\usage{
CASMAPenv
}
\description{
An environment to store a few global variables. Internal.
}
\keyword{internal}
