% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{checkIsBoolean}
\alias{checkIsBoolean}
\title{Check if a variable is boolean or not}
\usage{
checkIsBoolean(var, name)
}
\arguments{
\item{var}{The variable to be checked (if boolean).}

\item{name}{The name of the variable to appear in any error message.}
}
\value{
If not boolean (or \code{NA}), throws error.
        If \code{NA}, return \code{FALSE}. Otherwise return 
        boolean value of \code{var}.
}
\description{
Checks if a variable is boolean, if not throws an error, otherwise
returns boolean.
}
\keyword{internal}
