% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getExampleCovariatesFilename}
\alias{getExampleCovariatesFilename}
\title{Get the path to the example covariates file for regionGWAS mode}
\format{A single column vector of 100 labels, each of which
        is \code{0} or \code{1} (same format as labels file).}
\usage{
getExampleCovariatesFilename()
}
\description{
Path to \code{CASMAP_example_covariates_1.txt} in \code{inst/extdata}.
The covariates categories for the data set  
\code{CASMAP_example_data_1.txt}, the path to which is given by 
\code{getExampleDataFilename}.
}
\details{
Path to the file containing the labels, for reading in to
         CASMAP object using the \code{readFiles} function.
}
\examples{
covfile <- getExampleCovariatesFilename()
}
\seealso{
\code{getExampleDataFilename}, 
         \code{getExampleLabelsFilename}
}
