% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getExampleDataFilename}
\alias{getExampleDataFilename}
\title{Get the path to the example data file for regionGWAS mode}
\format{A matrix of \code{0}s and \code{1}s, with 1000 rows (features) 
        and 100 columns
        (samples). In other words, each column is a sample, and each sample
        has 1000 binary features.}
\usage{
getExampleDataFilename()
}
\description{
Path to \code{CASMAP_example_data_1.txt} in \code{inst/extdata}.
A dataset containing binary samples for the regionGWAS method.
There are accompanying labels and covariates dataset.
}
\details{
Path to the file containing the data, for reading in to
         CASMAP object using the \code{readFiles} function.
         Note that the significant region is \code{[99, 102]}.
}
\examples{
datafile <- getExampleDataFilename()
}
\seealso{
\code{getExampleLabelsFilename}, 
         \code{getExampleCovariatesFilename}
}
