% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getExampleLabelsFilename}
\alias{getExampleLabelsFilename}
\title{Get the path to the example labels file for regionGWAS mode}
\format{A single column of 100 labels, each of which is either \code{0}
        or \code{1}.}
\usage{
getExampleLabelsFilename()
}
\description{
Path to \code{CASMAP_example_labels_1.txt} in \code{inst/extdata}.
A dataset containing the binary labels for the data in the file 
\code{CASMAP_example_data_1.txt}, the path to which is given by 
\code{getExampleDataFilename}.
}
\details{
Path to the file containing the labels, for reading in to
         CASMAP object using the \code{readFiles} function.
}
\examples{
labelsfile <- getExampleLabelsFilename()
}
\seealso{
\code{getExampleDataFilename}, 
         \code{getExampleCovariatesFilename}
}
