% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{getExampleSignificantRegionsFilename}
\alias{getExampleSignificantRegionsFilename}
\title{Get the path to the example significant intervals file}
\usage{
getExampleSignificantRegionsFilename()
}
\description{
Path to \code{CASMAP_example_covariates_1.txt} in 
\code{inst/extdata}.
}
\examples{
sigregfile <- getExampleSignificantRegionsFilename()
}
\keyword{internal}
