% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getHigherOrderEpistasisString}
\alias{getHigherOrderEpistasisString}
\title{Gets the higherOrderEpistasis string}
\usage{
getHigherOrderEpistasisString()
}
\description{
A getter for the global \code{higherOrderEpistasis} value, a string
for the mode parameter.
}
\keyword{internal}
