% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMinModeLength}
\alias{getMinModeLength}
\alias{getMinModeLength}
\title{Gets the minModeLength}
\usage{
getMinModeLength()

getMinModeLength()
}
\description{
A getter for the global \code{minModeLength} value, a string
for the mode parameter.

Gets the minimum mode character length (should be 3)
}
\keyword{internal}
