% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRegionGWASString}
\alias{getRegionGWASString}
\title{Gets the regionGWAS string}
\usage{
getRegionGWASString()
}
\description{
A getter for the global \code{regionGWAS} value, a string
for the mode parameter.
}
\keyword{internal}
