% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isHigherOrderEpistasisString}
\alias{isHigherOrderEpistasisString}
\title{Checks if substring is part of higherOrderEpistasis}
\usage{
isHigherOrderEpistasisString(x)
}
\arguments{
\item{x}{The string which will be compared to 'higherOrderEpistasis'}
}
\value{
\code{TRUE} if the string is a substring of 'higherOrderEpistasis',
        otherwise returns \code{FALSE}.
}
\description{
Using grep to search through vector of strings
}
\details{
Uses \code{grep} to search for exact match.
}
\keyword{internal}
