% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{isInOpenInterval}
\alias{isInOpenInterval}
\title{A method to check value is numeric and in open interval}
\usage{
isInOpenInterval(x, lower = 0, upper = 1)
}
\arguments{
\item{x}{Value to be checked. Needs to be numeric.}

\item{lower}{Lower bound. Default value is \code{0}.}

\item{upper}{Upper bound. Default value is \code{1}.}
}
\value{
If numeric, and  strictly greater than \code{lower} and 
        strictly smaller than \code{upper}, then return \code{TRUE}. 
        Else return \code{FALSE}.
}
\description{
Checks if a value is numeric and strictly between two other values.
}
\keyword{internal}
