% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lib_pvals_significant_isets_write_to_file}
\alias{lib_pvals_significant_isets_write_to_file}
\title{Internal function}
\usage{
lib_pvals_significant_isets_write_to_file(inst, output_file)
}
\description{
Internal function
}
\keyword{internal}
