% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lib_read_covariates_file_fastcmh}
\alias{lib_read_covariates_file_fastcmh}
\title{Internal function}
\usage{
lib_read_covariates_file_fastcmh(inst, cov_filename)
}
\description{
Internal function
}
\keyword{internal}
