% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lib_read_plink_files_with_cov_facs}
\alias{lib_read_plink_files_with_cov_facs}
\title{Internal function}
\usage{
lib_read_plink_files_with_cov_facs(inst, base_filename, covfilename, encoding)
}
\description{
Internal function
}
\keyword{internal}
