% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lib_summary_write_to_file_fastcmh}
\alias{lib_summary_write_to_file_fastcmh}
\title{Internal function}
\usage{
lib_summary_write_to_file_fastcmh(inst, output_file)
}
\description{
Internal function
}
\keyword{internal}
