% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lib_write_eth_files_iset}
\alias{lib_write_eth_files_iset}
\title{Internal function}
\usage{
lib_write_eth_files_iset(inst, x_filename, y_filename)
}
\description{
Internal function
}
\keyword{internal}
