% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lib_write_eth_files_with_cov_fastcmh}
\alias{lib_write_eth_files_with_cov_fastcmh}
\title{Internal function}
\usage{
lib_write_eth_files_with_cov_fastcmh(inst, x_filename, y_filename, covfilename)
}
\description{
Internal function
}
\keyword{internal}
