% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\docType{class}
\name{m_SignificantFeaturesSearchWithCovariates-class}
\alias{m_SignificantFeaturesSearchWithCovariates-class}
\alias{m_SignificantFeaturesSearchWithCovariates}
\title{Internal class}
\description{
Internal class
}

\keyword{internal}
