% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\docType{class}
\name{m_SignificantIntervalSearchFais-class}
\alias{m_SignificantIntervalSearchFais-class}
\alias{m_SignificantIntervalSearchFais}
\title{Internal class}
\description{
An internal class.
}

\keyword{internal}
