% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{modeLengthErrorMessage}
\alias{modeLengthErrorMessage}
\title{Error message for mode, if too short}
\usage{
modeLengthErrorMessage()
}
\description{
Return the appropriate error message for incorrect mode input
}
\keyword{internal}
