% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\docType{class}
\name{SignificantIntervalSearchChi-class}
\alias{SignificantIntervalSearchChi-class}
\alias{SignificantIntervalSearchChi}
\title{Approximate fast significant interval search}
\description{
Class for approximate significant intervals search with Tarone correction for
bounding intermediate FWERs.
}

\keyword{internal}
