% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\docType{class}
\name{SignificantIntervalSearchExact-class}
\alias{SignificantIntervalSearchExact-class}
\alias{SignificantIntervalSearchExact}
\title{Internal class for search for significant regions}
\description{
Please use the \code{CASMAP} constructor.
}

\keyword{internal}
