% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\docType{class}
\name{SignificantIntervalSearchFastCmh-class}
\alias{SignificantIntervalSearchFastCmh-class}
\alias{SignificantIntervalSearchFastCmh}
\title{Fast significant interval search with categorical covariates}
\description{
Internal class, please use \code{CASMAP} constructor.
}

\keyword{internal}
