% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\docType{class}
\name{SignificantItemsetSearchFacs-class}
\alias{SignificantItemsetSearchFacs-class}
\alias{SignificantItemsetSearchFacs}
\title{Significant itemsets search with categorical covariates}
\description{
Internal class, please use \code{CASMAP} constructor.
}

\keyword{internal}
